updated tests to work after adding the token authentication system

+++ part4/tests/blog_api.test.js
@@ -4,6 +4,8 @@ const Blog = require('../models/blog')
 const app = require('../app')
 const api = supertest(app)
 
+const token = 'Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InJvb3QiLCJpZCI6IjYzYTgyZmEzN2QxN2E2YzAzMTFiN2NhMSIsImlhdCI6MTY3MjE1MTE4MH0.JVqPHb8TYOaTfPgJKmaCWlkthvTdhMHfB6c1Ty1rHdE'
+
 const initialBlogs = [
   {
     title: 'React patterns',
@@ -26,13 +28,13 @@ beforeEach(async () => {
 
 describe('tests for when there is some data stored in database', () => {
   test('number of blogs is correct', async () => {
-    const response = await api.get('/api/blogs')
+    const response = await api.get('/api/blogs').set('Authorization', token)
   
     expect(response.body).toHaveLength(initialBlogs.length)
   })
   
   test('the default _id in mongoose is changed to just id', async () => {
-    const response = await api.get('/api/blogs')
+    const response = await api.get('/api/blogs').set('Authorization', token)
   
     expect(response.body[0].id).toBeDefined()
   }, 100000)
@@ -55,21 +57,20 @@ describe('tests for when there is some data stored in database', () => {
       likes: 12
     }
   
-    await api.post('/api/blogs').send(newBlog).expect(400)
+    await api.post('/api/blogs').set('Authorization', token).send(newBlog).expect(400)
   })
 })
 
 describe('testing all REST API requests', () => {
   test('testing POST request', async () => {
     const newBlog = {
-      id: '5a422b3a1b54a676234d17f9',
       title: 'Canonical string reduction',
       author: 'Edsger W. Dijkstra',
       url: 'http://www.cs.utexas.edu/~EWD/transcriptions/EWD08xx/EWD808.html',
       likes: 12
     }
   
-    await api.post('/api/blogs').send(newBlog).expect(201)
+    await api.post('/api/blogs').set('Authorization', token).send(newBlog).expect(201)
   
     const allBlogs = await Blog.find({})
   
@@ -80,7 +81,7 @@ describe('testing all REST API requests', () => {
     let allBlogs = await Blog.find({})
     const id = allBlogs[0]._id.toHexString()
 
-    await api.delete(`/api/blogs/${id}`).expect(204)
+    await api.delete(`/api/blogs/${id}`).set('Authorization', token).expect(204)
 
     allBlogs = await Blog.find({})
 
@@ -91,7 +92,7 @@ describe('testing all REST API requests', () => {
     let allBlogs = await Blog.find({})
     const id = allBlogs[0]._id.toHexString()
 
-    await api.put(`/api/blogs/${id}`).expect(200)
+    await api.put(`/api/blogs/${id}`).set('Authorization', token).expect(200)
   })
 })
